[ui] ImageGallery: reset the GridView index when a project file is opened
This commits ensures that the GridView's index is correctly updated
when the projet file changes. Prior to this commit, the GridView's index
remained unchanged despite the change of images in the gallery. This
meant that if the newly opened project file had images whose size was
different from the previous project's, no auto-fit was performed in the
2D Viewer to display the newly images correctly, and manually triggering
the auto-fit was failing, as the GridView was unaware of its own content
until a manual change of index was done.