[ui] ImageGallery: display the correct image in the Viewer after a removal
When an image is removed from the gallery, the adjacent one is
automatically selected in the GridView. However, as the index in the
GridView would not change (the adjacent image taking the place of the
removed one), the 2D viewer would not receive any update and no more image
would be displayed. Even if the index in the GridView remains the same,
the 2D viewer now is updated with the newly selected image.